create new class for category to enhance encapsulation and abstraction
