--- src/test/java/seedu/planpal/contacts/ContactCategory.java
@@ -1,4 +1,0 @@
-package seedu.planpal.contacts;
-
-public class ContactCategory {
-}
